commit halo75 and update qmk

+++ keyboards/nuphy/air60_v2/ansi/rf.c
@@ -48,8 +48,8 @@ extern host_driver_t  *m_host_driver;
 extern uint8_t         host_mode;
 extern uint8_t         rf_blink_cnt;
 extern uint16_t        rf_link_show_time;
-extern uint32_t        rf_linking_time;
-extern uint32_t        no_act_time;
+extern uint16_t        rf_linking_time;
+extern uint16_t        no_act_time;
 extern bool            f_send_channel;
 extern bool            f_dial_sw_init_ok;
